refactor: split lifecycle handlers and open link handlers (#11852)
* refactor: handlers split

* refactor: handlers split

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut

* refactor: ut
