+++ packages/vscode-extension/src/debug/taskTerminal/lifecycleTaskTerminal.ts
@@ -9,7 +9,7 @@ import * as vscode from "vscode";
 import { err, FxError, ok, Result, Stage, Void } from "@microsoft/teamsfx-api";
 import { Correlator, TaskDefaultValue } from "@microsoft/teamsfx-core";
 import { workspaceUri } from "../../globalVariables";
-import { runCommand } from "../../handlers";
+import { runCommand } from "../../handlers/sharedOpts";
 import { TelemetryEvent, TelemetryProperty } from "../../telemetry/extTelemetryEvents";
 import { getLocalDebugSession } from "../common/localDebugSession";
 import { localTelemetryReporter, maskValue } from "../localTelemetryReporter";
