+++ packages/vscode-extension/src/extension.ts
@@ -3,9 +3,6 @@
 
 "use strict";
 
-import * as semver from "semver";
-import * as vscode from "vscode";
-
 import {
   AppPackageFolderName,
   BuildFolderName,
@@ -16,12 +13,14 @@ import {
 } from "@microsoft/teamsfx-api";
 import {
   AuthSvcScopes,
-  Correlator,
   FeatureFlags as CoreFeatureFlags,
+  Correlator,
   VersionState,
   featureFlagManager,
   teamsDevPortalClient,
 } from "@microsoft/teamsfx-core";
+import * as semver from "semver";
+import * as vscode from "vscode";
 
 import {
   CHAT_EXECUTE_COMMAND_ID,
@@ -81,7 +80,23 @@ import { checkSideloadingCallback } from "./handlers/checkSideloading";
 import * as copilotChatHandlers from "./handlers/copilotChatHandlers";
 import { debugInTestToolHandler } from "./handlers/debugInTestTool";
 import { downloadSampleApp } from "./handlers/downloadSample";
+import { deployHandler, provisionHandler, publishHandler } from "./handlers/lifecycleHandlers";
 import * as officeDevHandlers from "./handlers/officeDevHandlers";
+import {
+  openAccountLinkHandler,
+  openAppManagement,
+  openAzureAccountHandler,
+  openBotManagement,
+  openDevelopmentLinkHandler,
+  openDocumentHandler,
+  openDocumentLinkHandler,
+  openEnvLinkHandler,
+  openHelpFeedbackLinkHandler,
+  openLifecycleLinkHandler,
+  openM365AccountHandler,
+  openReportIssues,
+  openWelcomeHandler,
+} from "./handlers/openLinkHandlers";
 import { showOutputChannelHandler } from "./handlers/showOutputChannel";
 import { createProjectFromWalkthroughHandler } from "./handlers/walkthrough";
 import { ManifestTemplateHoverProvider } from "./hoverProvider";
@@ -306,7 +321,7 @@ function registerActivateCommands(context: vscode.ExtensionContext) {
   context.subscriptions.push(openLifecycleTreeview);
 
   // Documentation
-  registerInCommandController(context, CommandKeys.OpenDocument, handlers.openDocumentHandler);
+  registerInCommandController(context, CommandKeys.OpenDocument, openDocumentHandler);
 
   // README
   registerInCommandController(context, CommandKeys.OpenReadMe, handlers.openReadMeHandler);
@@ -315,7 +330,7 @@ function registerActivateCommands(context: vscode.ExtensionContext) {
   registerInCommandController(context, CommandKeys.OpenSamples, handlers.openSamplesHandler);
 
   // Quick start
-  registerInCommandController(context, CommandKeys.OpenWelcome, handlers.openWelcomeHandler);
+  registerInCommandController(context, CommandKeys.OpenWelcome, openWelcomeHandler);
   registerInCommandController(
     context,
     CommandKeys.BuildIntelligentAppsWalkthrough,
@@ -518,12 +533,7 @@ function registerTreeViewCommandsInDevelopment(context: vscode.ExtensionContext)
 
 function registerTreeViewCommandsInLifecycle(context: vscode.ExtensionContext) {
   // Provision in the cloud
-  registerInCommandController(
-    context,
-    CommandKeys.Provision,
-    handlers.provisionHandler,
-    "provision"
-  );
+  registerInCommandController(context, CommandKeys.Provision, provisionHandler, "provision");
 
   // Zip Teams metadata package
   registerInCommandController(
@@ -534,10 +544,10 @@ function registerTreeViewCommandsInLifecycle(context: vscode.ExtensionContext) {
   );
 
   // Deploy to the cloud
-  registerInCommandController(context, CommandKeys.Deploy, handlers.deployHandler, "deploy");
+  registerInCommandController(context, CommandKeys.Deploy, deployHandler, "deploy");
 
   // Publish to Teams
-  registerInCommandController(context, CommandKeys.Publish, handlers.publishHandler, "publish");
+  registerInCommandController(context, CommandKeys.Publish, publishHandler, "publish");
 
   // Publish in Developer Portal
   registerInCommandController(
@@ -548,16 +558,12 @@ function registerTreeViewCommandsInLifecycle(context: vscode.ExtensionContext) {
   );
 
   // Developer Portal for Teams
-  registerInCommandController(
-    context,
-    "fx-extension.openAppManagement",
-    handlers.openAppManagement
-  );
+  registerInCommandController(context, "fx-extension.openAppManagement", openAppManagement);
 }
 
 function registerTreeViewCommandsInHelper(context: vscode.ExtensionContext) {
   // Report issues on GitHub
-  registerInCommandController(context, "fx-extension.openReportIssues", handlers.openReportIssues);
+  registerInCommandController(context, "fx-extension.openReportIssues", openReportIssues);
 }
 
 /**
@@ -591,7 +597,7 @@ function registerTeamsFxCommands(context: vscode.ExtensionContext) {
 
   const openBotManagementCmd = vscode.commands.registerCommand(
     "fx-extension.openBotManagement",
-    (...args) => Correlator.run(handlers.openBotManagement, args)
+    (...args) => Correlator.run(openBotManagement, args)
   );
   context.subscriptions.push(openBotManagementCmd);
 
@@ -643,7 +649,7 @@ function registerMenuCommands(context: vscode.ExtensionContext) {
 
   const azureAccountSettingsCmd = vscode.commands.registerCommand(
     "fx-extension.azureAccountSettings",
-    () => Correlator.run(handlers.openAzureAccountHandler)
+    () => Correlator.run(openAzureAccountHandler)
   );
   context.subscriptions.push(azureAccountSettingsCmd);
 
@@ -685,50 +691,45 @@ function registerMenuCommands(context: vscode.ExtensionContext) {
 
   const m365AccountSettingsCmd = vscode.commands.registerCommand(
     "fx-extension.m365AccountSettings",
-    () => Correlator.run(handlers.openM365AccountHandler)
+    () => Correlator.run(openM365AccountHandler)
   );
   context.subscriptions.push(m365AccountSettingsCmd);
 
   const openAccountLinkCmd = vscode.commands.registerCommand(
     "fx-extension.openAccountLink",
-    (...args) =>
-      Correlator.run(handlers.openAccountLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
+    (...args) => Correlator.run(openAccountLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
   );
   context.subscriptions.push(openAccountLinkCmd);
 
   const openLifecycleLinkCmd = vscode.commands.registerCommand(
     "fx-extension.openLifecycleLink",
     (...args) =>
-      Correlator.run(handlers.openLifecycleLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
+      Correlator.run(openLifecycleLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
   );
   context.subscriptions.push(openLifecycleLinkCmd);
 
   const openDevelopmentLinkCmd = vscode.commands.registerCommand(
     "fx-extension.openDevelopmentLink",
     (...args) =>
-      Correlator.run(handlers.openDevelopmentLinkHandler, [
-        TelemetryTriggerFrom.ViewTitleNavigation,
-      ])
+      Correlator.run(openDevelopmentLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
   );
   context.subscriptions.push(openDevelopmentLinkCmd);
 
   const openEnvLinkCmd = vscode.commands.registerCommand("fx-extension.openEnvLink", (...args) =>
-    Correlator.run(handlers.openEnvLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
+    Correlator.run(openEnvLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
   );
   context.subscriptions.push(openEnvLinkCmd);
 
   const openHelpFeedbackLinkCmd = vscode.commands.registerCommand(
     "fx-extension.openHelpFeedbackLink",
     (...args) =>
-      Correlator.run(handlers.openHelpFeedbackLinkHandler, [
-        TelemetryTriggerFrom.ViewTitleNavigation,
-      ])
+      Correlator.run(openHelpFeedbackLinkHandler, [TelemetryTriggerFrom.ViewTitleNavigation])
   );
   context.subscriptions.push(openHelpFeedbackLinkCmd);
 
   const openDocumentLinkCmd = vscode.commands.registerCommand(
     "fx-extension.openDocumentLink",
-    (...args) => Correlator.run(handlers.openDocumentLinkHandler, args)
+    (...args) => Correlator.run(openDocumentLinkHandler, args)
   );
   context.subscriptions.push(openDocumentLinkCmd);
