+++ packages/vscode-extension/src/handlers/lifecycleHandlers.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+import { FxError, Result, Stage } from "@microsoft/teamsfx-api";
+import { isUserCancelError } from "@microsoft/teamsfx-core";
+import { ExtTelemetry } from "../telemetry/extTelemetry";
+import { TelemetryEvent } from "../telemetry/extTelemetryEvents";
+import envTreeProviderInstance from "../treeview/environmentTreeViewProvider";
+import { getTriggerFromProperty } from "../utils/telemetryUtils";
+import { runCommand } from "./sharedOpts";
+
+export async function provisionHandler(...args: unknown[]): Promise<Result<unknown, FxError>> {
+  ExtTelemetry.sendTelemetryEvent(TelemetryEvent.ProvisionStart, getTriggerFromProperty(args));
+  const result = await runCommand(Stage.provision);
+  if (result.isErr() && isUserCancelError(result.error)) {
+    return result;
+  } else {
+    // refresh env tree except provision cancelled.
+    await envTreeProviderInstance.reloadEnvironments();
+    return result;
+  }
+}
+
+export async function deployHandler(...args: unknown[]): Promise<Result<null, FxError>> {
+  ExtTelemetry.sendTelemetryEvent(TelemetryEvent.DeployStart, getTriggerFromProperty(args));
+  return await runCommand(Stage.deploy);
+}
+
+export async function publishHandler(...args: unknown[]): Promise<Result<null, FxError>> {
+  ExtTelemetry.sendTelemetryEvent(TelemetryEvent.PublishStart, getTriggerFromProperty(args));
+  return await runCommand(Stage.publish);
+}
