+++ packages/vscode-extension/test/extension/controls/openWelcomePage.test.ts
@@ -1,9 +1,7 @@
 import * as chai from "chai";
 import * as sinon from "sinon";
 import * as vscode from "vscode";
-
 import * as globalState from "@microsoft/teamsfx-core/build/common/globalState";
-
 import { openWelcomePageAfterExtensionInstallation } from "../../../src/controls/openWelcomePage";
 import { ExtTelemetry } from "../../../src/telemetry/extTelemetry";
