+++ packages/vscode-extension/test/mocks/mockCore.ts
@@ -31,7 +31,9 @@ export class MockCore {
   async deployAadManifest(inputs: Inputs): Promise<Result<undefined, FxError>> {
     return ok(undefined);
   }
-
+  async deployTeamsManifest(inputs: Inputs): Promise<Result<undefined, FxError>> {
+    return ok(undefined);
+  }
   async deployArtifacts(inputs: Inputs): Promise<Result<undefined, FxError>> {
     return ok(undefined);
   }
@@ -43,7 +45,9 @@ export class MockCore {
   async publishApplication(inputs: Inputs): Promise<Result<undefined, FxError>> {
     return ok(undefined);
   }
-
+  async createAppPackage(inputs: Inputs): Promise<Result<undefined, FxError>> {
+    return ok(undefined);
+  }
   async executeUserTask(func: Func, inputs: Inputs): Promise<Result<any, FxError>> {
     return ok("");
   }
